// sop-ui/src/utils/types.ts
export type FormInitialValues = {
  name: string;
  email: string;
  age: string;
  highestEducation: string;
  institute: string;
  study: string;
  workExperience: string;
  instituteInCanada: string;
  programInCanada: string;
  countryApplyingFrom: string;
  futureGoals: string;
  englishListening: number;
  englishReading: number;
  englishSpeaking: number;
  englishWriting: number;
  paidFirstYearTuition: string;
  tuitionFee: number;
  didGIC: string;
  gicAmount: number;
};